Solving of all suggestion specified in the pull request form problem 7
