+++ src/com/warren/projecteuler/Problem10.java
@@ -1,5 +1,7 @@
 package com.warren.projecteuler;
 
+import com.warren.Utils;
+
 import java.math.BigInteger;
 
 public class Problem10 {
@@ -8,16 +10,15 @@ public class Problem10 {
     * Find the sum of all the primes below two million.
     * */
 
-    public static void sumPrimeBelowTo(int maxValueBelow) {
-
-        BigInteger sumPrimeNumber = BigInteger.valueOf(0);
-        for (int number = 2; number <= maxValueBelow; number++){
-            if (Problem7.isPrime(number)){
-                sumPrimeNumber = sumPrimeNumber.add(BigInteger.valueOf(number));
+    public static void sumOfPrimeNumbersBelow(int number){
+        BigInteger result = BigInteger.ZERO;
+        for (int i = 2; i <= number; i++){
+            if (Utils.isPrime(i)){
+                result = result.add(BigInteger.valueOf(i));
             }
         }
 
-        System.out.println("Problem 10 : The sum of all the primes below to "+maxValueBelow+" is : "+sumPrimeNumber);
+        System.out.println("Problem 10 : The sum of all the primes below to "+number+" is : "+result);
         System.out.println("***********************************************");
     }
 }
